app/requirements_backup.txt
openai
streamlit
python-dotenv
numexpr
langchainhub
db-dtypes
langchain
pydantic==1.10.13
pymysql
pandas
docx
python-docx
langchain==0.1.16
langchain-community==0.0.34
gpt4all==1.0.8
chromadb==0.4.7
llama-cpp-python==0.1.81
urllib3==2.0.4
PyMuPDF==1.23.5
python-dotenv==1.0.0W
unstructured==0.10.8
extract-msg==0.45.0
tabulate==0.9.0
pandoc==2.3
pypandoc==1.11
tqdm==4.66.1
sentence_transformers==2.2.2
huggingface_hub